refactor: move stat methods to trait
